requirements.txt
flask
numpy
cassandra-driver